refactor datasets functions to unify VV and map
