+++ apps/fishing-map/features/datasets/datasets.utils.ts
@@ -144,40 +144,6 @@ export const getDatasetsInDataviews = (
   )
 }
 
-export const getRelatedDatasetByType = (
-  dataset?: Dataset,
-  datasetType?: DatasetTypes,
-  fullDatasetAllowed = false
-) => {
-  if (fullDatasetAllowed) {
-    const fullDataset = dataset?.relatedDatasets?.find(
-      (relatedDataset) =>
-        relatedDataset.type === datasetType && relatedDataset.id.startsWith(FULL_SUFIX)
-    )
-    if (fullDataset) {
-      return fullDataset
-    }
-  }
-  return dataset?.relatedDatasets?.find((relatedDataset) => relatedDataset.type === datasetType)
-}
-
-export const getRelatedDatasetsByType = (
-  dataset?: Dataset,
-  datasetType?: DatasetTypes,
-  fullDatasetAllowed = false
-) => {
-  if (fullDatasetAllowed) {
-    const fullDataset = dataset?.relatedDatasets?.filter(
-      (relatedDataset) =>
-        relatedDataset.type === datasetType && relatedDataset.id.startsWith(FULL_SUFIX)
-    )
-    if (fullDataset.length > 0) {
-      return fullDataset
-    }
-  }
-  return dataset?.relatedDatasets?.filter((relatedDataset) => relatedDataset.type === datasetType)
-}
-
 export const getActiveDatasetsInActivityDataviews = (
   dataviews: UrlDataviewInstance<GeneratorType>[]
 ) => {
