+++ apps/vessel-history/features/datasets/datasets.selectors.ts
@@ -1,8 +1,7 @@
 import { createSelector } from '@reduxjs/toolkit'
 import { uniqBy } from 'lodash'
 import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
-import { Dataset, DatasetTypes } from '@globalfishingwatch/api-types'
-import { FULL_SUFIX } from 'data/config'
+import { DatasetTypes } from '@globalfishingwatch/api-types'
 import { selectDatasets } from './datasets.slice'
 
 export const getDatasetsByDataview = (dataview: UrlDataviewInstance) =>
@@ -44,24 +43,3 @@ export const selectTracksDatasets = createSelector(
     return datasets
   }
 )
-
-export const getRelatedDatasetByType = (
-  dataset?: Dataset,
-  datasetType?: DatasetTypes,
-  fullDatasetAllowed = false
-) => {
-  if (fullDatasetAllowed) {
-    const fullDataset = dataset?.relatedDatasets?.find(
-      (relatedDataset) =>
-        relatedDataset.type === datasetType && relatedDataset.id.startsWith(FULL_SUFIX)
-    )
-    if (fullDataset) {
-      return fullDataset
-    }
-  }
-  return dataset?.relatedDatasets?.find((relatedDataset) => relatedDataset.type === datasetType)
-}
-
-export const getRelatedDatasetsByType = (dataset?: Dataset, datasetType?: DatasetTypes) => {
-  return dataset?.relatedDatasets?.filter((relatedDataset) => relatedDataset.type === datasetType)
-}
