+++ apps/vessel-history/features/dataviews/dataviews.utils.ts
@@ -4,30 +4,24 @@ import {
   DataviewDatasetConfig,
   DataviewInstance,
   EndpointId,
+  VesselInstanceDatasets,
 } from '@globalfishingwatch/api-types'
 import { GeneratorType } from '@globalfishingwatch/layer-composer'
 import { UrlDataviewInstance } from '@globalfishingwatch/dataviews-client'
-import { AppDispatch, RootState } from 'store'
+import { AppDispatch } from 'store'
 import { fetchDatasetsByIdsThunk } from 'features/datasets/datasets.slice'
-import { selectUserLogged } from 'features/user/user.slice'
 import { fetchDataviewsByIdsThunk } from './dataviews.slice'
 import {
   dataviewInstances,
   vesselDataviewIds,
   DEFAULT_TRACK_COLOR,
   MAP_BACKGROUND_COLOR,
-  TEMPLATE_VESSEL_DATAVIEW_ID,
 } from './dataviews.config'
 
 // used in workspaces with encounter events layers
 export const VESSEL_LAYER_PREFIX = 'vessel-'
 export const VESSEL_DATAVIEW_INSTANCE_PREFIX = 'vessel-'
 
-type VesselInstanceDatasets = {
-  trackDatasetId: string
-  infoDatasetId?: string
-  eventsDatasetsId?: string[]
-}
 export const getVesselDataviewInstanceId = (vesselId: string) => `${VESSEL_LAYER_PREFIX}${vesselId}`
 
 export const getVesselDataviewInstanceFactory =
@@ -114,48 +108,4 @@ export const initializeDataviews = async (dispatch: AppDispatch) => {
     const datasets = getDatasetByDataview(dataviews)
     dispatch(fetchDatasetsByIdsThunk(datasets))
   }
-}
-
-export const getVesselDataviewInstanceDatasetConfig = (
-  vesselId: string,
-  { trackDatasetId, infoDatasetId, eventsDatasetsId }: VesselInstanceDatasets
-) => {
-  const datasetsConfig: DataviewDatasetConfig[] = []
-  if (infoDatasetId) {
-    datasetsConfig.push({
-      datasetId: infoDatasetId,
-      params: [{ id: 'vesselId', value: vesselId }],
-      endpoint: EndpointId.Vessel,
-    })
-  }
-  if (trackDatasetId) {
-    datasetsConfig.push({
-      datasetId: trackDatasetId,
-      params: [{ id: 'vesselId', value: vesselId }],
-      endpoint: EndpointId.Tracks,
-    })
-  }
-  if (eventsDatasetsId) {
-    eventsDatasetsId.forEach((eventDatasetId) => {
-      datasetsConfig.push({
-        datasetId: eventDatasetId,
-        query: [{ id: 'vessels', value: vesselId }],
-        params: [],
-        endpoint: EndpointId.Events,
-      })
-    })
-  }
-  return datasetsConfig
-}
-
-export const getVesselDataviewInstance = (
-  vessel: { id: string },
-  datasets: VesselInstanceDatasets
-): DataviewInstance<GeneratorType> => {
-  const vesselDataviewInstance = {
-    id: `${VESSEL_DATAVIEW_INSTANCE_PREFIX}${vessel.id}`,
-    dataviewId: TEMPLATE_VESSEL_DATAVIEW_ID,
-    datasetsConfig: getVesselDataviewInstanceDatasetConfig(vessel.id, datasets),
-  }
-  return vesselDataviewInstance
 }
