+++ libs/dataviews-client/src/index.ts
@@ -1,6 +1,8 @@
 export * from './dataviews-client'
+export * from './resolve-datasets'
 export * from './resolve-dataviews'
 export * from './resolve-dataviews-generators'
 export * from './resolve-endpoint'
 export * from './resources'
 export * from './url-workspace'
+export * from './utils'
