+++ libs/dataviews-client/src/resolve-datasets.ts
@@ -1,0 +1,38 @@
+import { Dataset, DatasetTypes } from '@globalfishingwatch/api-types'
+
+export const FULL_SUFIX = 'full'
+
+export const getRelatedDatasetByType = (
+  dataset?: Dataset,
+  datasetType?: DatasetTypes,
+  fullDatasetAllowed = false
+) => {
+  if (fullDatasetAllowed) {
+    const fullDataset = dataset?.relatedDatasets?.find(
+      (relatedDataset) =>
+        relatedDataset.type === datasetType && relatedDataset.id.startsWith(FULL_SUFIX)
+    )
+    if (fullDataset) {
+      return fullDataset
+    }
+  }
+  return dataset?.relatedDatasets?.find((relatedDataset) => relatedDataset.type === datasetType)
+}
+
+export const getRelatedDatasetsByType = (
+  dataset?: Dataset,
+  datasetType?: DatasetTypes,
+  fullDatasetAllowed = false
+) => {
+  if (fullDatasetAllowed) {
+    const fullDataset =
+      dataset?.relatedDatasets?.filter(
+        (relatedDataset) =>
+          relatedDataset.type === datasetType && relatedDataset.id.startsWith(FULL_SUFIX)
+      ) || []
+    if (fullDataset.length > 0) {
+      return fullDataset
+    }
+  }
+  return dataset?.relatedDatasets?.filter((relatedDataset) => relatedDataset.type === datasetType)
+}
